refactor(lumi-survey): extract types and simplify DockPanel prop forwarding
